Genauere Returnwerte

+++ AvorionServer.py
@@ -1,31 +1,37 @@
 import datetime
 import settings
 import valve.rcon
+import os.path
 
 def backup():
 	"""throws
 	"""
 		
 	"""returns
-	true/false
+	0	-	All ok
+	-2	-	Server could not be stopped
+	-3	-	Server could not be started
 	"""
 	
 	#Stop the Server
-	stop()
-	
-	#Create time and backupname
-	#Time cheatsheet: http://strftime.org/
-	dtNow	= datetime.datetime.now()
-	strNow	= dtNow.strftime('%Y-%m-%d_%H-%M-%S')
-	strBak	= settings.BackupPath + settings.BackupName +"_"+ strNow +".tgz"
-	
-	#Run backup
-	with tarfile.open(strBak, "w:gz") as tar:
-		tar.add(settings.ServerPath, arcname=os.path.basename(settings.ServerPath))
-	
-	
-	start()
-	return true
+	if stop():
+		#Create time and backupname
+		#Time cheatsheet: http://strftime.org/
+		dtNow	= datetime.datetime.now()
+		strNow	= dtNow.strftime('%Y-%m-%d_%H-%M-%S')
+		strBak	= settings.BackupPath + settings.BackupName +"_"+ strNow +".tgz"
+		
+		#Run backup
+		with tarfile.open(strBak, "w:gz") as tar:
+			tar.add(settings.ServerPath, arcname=os.path.basename(settings.ServerPath))
+		
+		
+		if start():
+			return 0
+		else
+			return -3
+	else
+		return -2
 
 def runRcon(strCommand):
     """throws
@@ -51,30 +57,40 @@ def stop():
     """
     
     """returns
-	string
-	false
+	0	-	all OK
+	-1	-	Lockfile could not be removed
 	"""
 	
-    strReturn = runRcon("/stop")
+	if os.path.isfile(settings.LockFile):
+		strReturn = runRcon("/stop")
     
-    #Delete the Lockfile
-    try:
-		os.remove(settings.LockFile)
-	catch OSError:
-		return False
-
-    return strReturn
+		#Delete the Lockfile
+		try:
+			os.remove(settings.LockFile)
+		catch OSError:
+			return -1
+		return 0
+	else
+		return 0
 
 def start():
+	"""returns
+	0	-	All OK
+	-1	-	Server already running / Lockfile exists
+	-2	-	Lockfile could not be created
+	"""
     #do some starting stuff
     
+    if os.path.isfile(settings.LockFile):
+		return -1
+	
     #Create a Lockfile
     try:
 		os.open(settings.LockFile,'w')
 	catch OSError:
-		return False 
+		return -2 
 		
-    return False
+    return 0
 
 def save():
     """throws
